✨: Collected Card View 삭제 확인
